+++ app/server.rb
@@ -73,19 +73,19 @@ class Chitter < Sinatra::Base
     end
   end
 
+  delete '/sessions' do
+    flash[:notice] = "Goodbye!"
+    session[:user_id] = nil
+    redirect to('/')
+  end
+
+
 
   helpers do
     def current_user
       @current_user ||=User.get(session[:user_id]) if session[:user_id]
     end
   end
-
-
-
-
-
-
-
 
   # start the server if ruby file executed directly
   run! if app_file == $0
